modify/correct Integral.is_number
isyms should start with the free symbols of the integrand, not all
symbols.

A 2nd non-heroic test for a case where a free symbol will not
affect evaluation was added to is_number and to evalf for integrals.
If the variable of integration does not appear in the free symbols
then only the difference in the limits is of significance, e.g.

Integral(x, (y, d, d + 1)) will introduce a linear factor of y
and the value of d is not significant. This integral is the same
as Integral(x, (y, 0, 1)).